+++ src/data/midi/message/mod.rs
@@ -1,14 +1,16 @@
 //! Enum representing the MIDI message.
 
+pub mod channel;
 pub mod control_function;
+pub mod notes;
 pub mod raw;
 
 use crate::data::byte::from_traits::FromClamped;
 use crate::data::byte::u7::U7;
-use crate::data::midi::channel::Channel;
+use crate::data::midi::message::channel::Channel;
 use crate::data::midi::message::control_function::ControlFunction;
+use crate::data::midi::message::notes::Note;
 use crate::data::midi::message::raw::{Payload, Raw};
-use crate::data::midi::notes::Note;
 use crate::packet::event_packet::MidiPacketParsingError;
 use core::convert::TryFrom;